Nano Maze for SHIELD
A SHIELD Optimised Port of Maze, aka Nano Maze or NanoMaze by Jesusthekiller, to SHIELD.

Nano Maze's successor, 3DMaze/Maze3D, is currently broken and there are no efforts ongoing to fix this. Because of this, it is unlikely we will see Maze3D on SHIELD. Anyway, enough about Maze3D, here's some info on NanoMaze:

• Maps can be up to 99x99
• Full Colour Support in 51x19
• 2D Bird's Eye Perspective
• Timed, so you can see your best time
And more!

We could hack it to have infinite sized mazes but it'd eventually crash, should we do that? I dunno, let's stick with 99x99 as max for now.